Move hooks de conexão com o mongodb/opac para arquivo adequado
